# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ENUMERA API",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "name": "ENUMERA",
            "url": "https://enumera.tech",
            "email": "enumera@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/evaluations-lecons": {
            "get": {
                "tags": [
                    "Evaluations leçons"
                ],
                "summary": "Liste des evaluations",
                "description": "Retourne la liste des evaluations",
                "operationId": "4997cebef99b6319c6b4b67480a94146",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Evaluations leçons"
                ],
                "summary": "Création d'une evaluation",
                "description": "Crée une nouvelle evaluation et retourne la evaluation créée",
                "operationId": "d29da88a1710a6f8b3af92887b2e1956",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "abreviation",
                                    "description",
                                    "lecon"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Intitulé de l'evaluation"
                                    },
                                    "abreviation": {
                                        "type": "string",
                                        "example": "Intitulé de l'evaluation"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "courte description"
                                    },
                                    "lecon": {
                                        "type": "string",
                                        "example": "Slug de la leçon"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Evaluation créée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/evaluations-lecons/{slug}": {
            "get": {
                "tags": [
                    "Evaluations leçons"
                ],
                "summary": "Récupère une evaluation par son slug",
                "description": "Retourne une evaluation",
                "operationId": "0d74701e30d03fa5d9a84a01a15d7483",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug de la evaluation à récupérer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Evaluations leçons"
                ],
                "summary": "Modification d'une evaluation",
                "description": "Modifie une evaluation et retourne la evaluation modifiée",
                "operationId": "8c123a369d2026fe502ed675bc42d438",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug de la evaluation à modifiée",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "abreviation",
                                    "description",
                                    "lecon"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Intitulé de l'evaluation"
                                    },
                                    "abreviation": {
                                        "type": "string",
                                        "example": "Intitulé de l'evaluation"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "courte description"
                                    },
                                    "lecon": {
                                        "type": "string",
                                        "example": "Slug de la leçon"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "evaluation modifiée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "evaluation non trouvée"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Evaluations leçons"
                ],
                "summary": "Suppression d'une evaluation par son slug",
                "description": "Retourne la evaluation supprimée",
                "operationId": "f97a12a0fb0df5da430ad978e5c3826e",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug de la evaluation à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "evaluation supprimée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "evaluation non trouvée"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/evaluations-lecons/import": {
            "post": {
                "tags": [
                    "Evaluations leçons"
                ],
                "summary": "Création d'une evaluation",
                "description": "Importe une liste de qcm",
                "operationId": "677fa034e53017fc7b894ecbe35d1bec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question",
                                    "choix",
                                    "type",
                                    "lecon"
                                ],
                                "properties": {
                                    "lecon": {
                                        "type": "string",
                                        "example": "Slug de la leçon"
                                    },
                                    "qcm": {
                                        "type": "string",
                                        "example": "fichier excel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Evaluation créée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}